+++ Solar/source/Solar/Layers/ImGui/ImGuiLayer.cpp
@@ -1,8 +1,8 @@
 ﻿#include "slpch.h"
 
-#include "imgui.h"
-#include "Platform/OpenGl/ImGuiOpenGLRenderer.h"
-#include "Platform/OpenGl/ImGuiGLFWRenderer.h"
+#include <imgui.h>
+#include <backends/imgui_impl_opengl3.h>
+#include <backends/imgui_impl_glfw.h>
 #include <glad/glad.h>
 
 #include "ImGuiLayer.h"
@@ -77,7 +77,7 @@ namespace Solar
    {
         ImGuiIO& io = ImGui::GetIO();
         const Application& app = Application::Get();
-        io.DisplaySize = ImVec2(app.GetWindow().GetWidth(), app.GetWindow().GetHeight());
+        io.DisplaySize = ImVec2(static_cast<float>(app.GetWindow().GetWidth()), static_cast<float>(app.GetWindow().GetHeight()));
 
          ImGui::Render();
          ImGui_ImplOpenGL3_RenderDrawData(ImGui::GetDrawData());
